make a utilities folder with plotting scripts and time series function for creating model initialization files; left MSSM model (written in Matlab and R) in the tools folder, but moved the csv file needed for the notebook to that folder
