--- Tools/Multivariate_Sea_Storm_Model/.Rapp.history
@@ -1,50 +1,0 @@
-library(CDVine)#
-library(R.matlab)#
-#library(VineCopula)#
-#
-# user inputs#
-U <- read.table("U_mssmVCR.txt")   # N x d data matrix (with uniform margins)#
-nSimNum <- read.table("Inputs_mssm.txt")   # 10000
-setwd("/Users/KatherineAnardeWheels/PyCharmProjects/Barrier3D/Tools/Multivariate_Sea_Storm_Model")
-library(CDVine)#
-library(R.matlab)
-U <- read.table("U_mssmVCR.txt")   # N x d data matrix (with uniform margins)#
-nSimNum <- read.table("Inputs_mssm.txt")   # 10000
-U <- read.table("U_mssm.txt")   # N x d data matrix (with uniform margins)#
-nSimNum <- read.table("Inputs_mssm.txt")   # 10000
-nSimNum
-# copula model parameters#
-type  <- "DVine";          # options are "CVine" or "DVine"#
-#
-# returns pair-copula families composing the C- or D-vine copula ($family), its parameters ($par and $par2) #
-Model <- CDVineCopSelect(U, familyset = NA, type = type, selectioncrit="AIC", indeptest=FALSE, level=0.05)#
-#
-# generate new data Usim from the copula#
-uSim  <- CDVineSim(nSimNum, family=Model$family, par=Model$par, Model$par2, type=type)#
-#
-# write outputs to a mat file#
-writeMat('Usim_mssmVCR-Dvine.mat', uSim = uSim, fixNames=TRUE, matVersion="5", onWrite=NULL, verbose=FALSE)#
-#
-# now do it again for the c-vine copula#
-type  <- "CVine";          # options are "CVine" or "DVine"#
-Model <- CDVineCopSelect(U, familyset = NA, type = type, selectioncrit="AIC", indeptest=FALSE, level=0.05)#
-uSim  <- CDVineSim(nSimNum, family=Model$family, par=Model$par, Model$par2, type=type)#
-writeMat('Usim_mssmVCR-Cvine.mat', uSim = uSim, fixNames=TRUE, matVersion="5", onWrite=NULL, verbose=FALSE)
-# copula model parameters#
-type  <- "DVine";          # options are "CVine" or "DVine"#
-#
-# returns pair-copula families composing the C- or D-vine copula ($family), its parameters ($par and $par2) #
-Model <- CDVineCopSelect(U, familyset = NA, type = type, selectioncrit="AIC", indeptest=FALSE, level=0.05)#
-#
-# generate new data Usim from the copula#
-uSim  <- CDVineSim(nSimNum, family=Model$family, par=Model$par, Model$par2, type=type)#
-#
-# write outputs to a mat file#
-writeMat('Usim_mssm-Dvine.mat', uSim = uSim, fixNames=TRUE, matVersion="5", onWrite=NULL, verbose=FALSE)#
-#
-# now do it again for the c-vine copula#
-type  <- "CVine";          # options are "CVine" or "DVine"#
-Model <- CDVineCopSelect(U, familyset = NA, type = type, selectioncrit="AIC", indeptest=FALSE, level=0.05)#
-uSim  <- CDVineSim(nSimNum, family=Model$family, par=Model$par, Model$par2, type=type)#
-writeMat('Usim_mssm-Cvine.mat', uSim = uSim, fixNames=TRUE, matVersion="5", onWrite=NULL, verbose=FALSE)
-uSim
